+++ src/converters/piral-gea/convert.ts
@@ -1,0 +1,29 @@
+import { createConverter } from './lib/converter';
+
+export interface HtmlComponent<TProps> {
+  component: {
+    mount(element: HTMLElement, props: TProps, ctx: any, locals: any): void;
+    update?(element: HTMLElement, props: TProps, ctx: any, locals: any): void;
+    unmount?(element: HTMLElement, locals: any): void;
+  };
+  type: 'html';
+}
+
+export interface GeaConverter {
+  (...params: Parameters<ReturnType<typeof createConverter>>): HtmlComponent<any>;
+}
+
+export function createGeaConverter(...params: Parameters<typeof createConverter>) {
+  const convert = createConverter(...params);
+  const Extension = convert.Extension;
+  const from: GeaConverter = (root) => ({
+    type: 'html',
+    component: convert(root),
+  });
+
+  return { from, Extension };
+}
+
+const { from: fromGea, Extension: GeaExtension } = createGeaConverter();
+
+export { fromGea, GeaExtension };